requirement.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
db-dtypes==1.2.0
google-api-core==2.19.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
joblib==1.3.2
numpy==1.26.3
oauthlib==3.2.2
packaging==24.0
pandas==2.2.0
pandas-gbq==0.23.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
setuptools==70.0.0
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
urllib3==2.2.1